backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
web3==6.11.0
pydantic==2.5.0
python-dotenv==1.0.0
eth-account==0.10.0
